requirements.txt
selenium==4.18.1
beautifulsoup4==4.12.3
requests==2.32.3
numpy==1.26.4
pandas==2.2.2
webdriver-manager==4.0.1
undetected-chromedriver==3.5.5
tqdm==4.67.0
xgboost==2.1.4
scikit-learn==1.6.1
tensorflow==2.15.0
matplotlib==3.10.0
python-telegram-bot==22.0
email-service==2.0.0
joblib==1.5.0
fastapi==0.115.12
uvicorn==0.34.2